refactor : api 호출 함수 분리
